+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 * Make clearer django migrate command in invoke tasks
+* `populatedb` command is now idempotent
 
 ## v0.5.0 - 2024-01-20